Add legacy media types
With the publication of the IronCore OCI image spec the old `v1alpha1`
media types were removed from the code. This was a backwards
incompatible change, as image layers created with the old media types
were no longer recognized.

This commit adds the media types back and marks them as legacy.
This will ensure backwards compatibility.